refractor check braces for blocks
